#491: bcopy's word is as wide as a pointer, and its casts stop truncating
`typedef int word' said four bytes on a machine whose registers are eight,
and the low-bit tests reached the pointers through `(int)' casts that threw
away the top half.  Both are now stated correctly: uintptr_t for the word,
uintptr_t for the casts.  Six -Wpointer-to-int-cast diagnoses go away by the
code being right rather than by the warning being silenced.

⚠️ It does NOT make the copy faster, and this issue was opened claiming it
would.  Measured, with objects built by the build's own command line -- the
new one's .text is 1371 bytes, byte-identical to the object that ships:

	bytes      old      new     A/B     A/A control
	   64       27       25   1.08x        1.00x
	  256       60       53   1.13x        1.00x
	 4096      748      746   1.00x        0.99x
	16384     2426     2423   1.00x        1.00x

The reason is that gcc at -O3 auto-vectorises both versions to movdqu/movups
and moves sixteen bytes at a time regardless of what the typedef says.  The
typedef shapes the loop that is written; the optimiser replaces it.  So "every
memcpy moves 32 bits per iteration" was true of the SOURCE and false of the
BINARY -- read from the file instead of asked of the build.

⚠️ Two instrument failures on the way to that table, both caught by controls
rather than by suspicion:

  - the first harness compiled with flags of my own choosing instead of the
    build's, and reported 2.9x.  Its .text was 558 bytes against the shipped
    1371, so it was timing code nobody runs.
  - an A/A control -- the same function in both positions -- reads 1.24x at 16
    bytes and 0.83x at 1024, so the harness is not honest at those sizes and
    their A/B numbers are worth nothing.  The rows are left in the table
    rather than dropped.

i386 does not build this file at all: it uses a 325-line SSE2 assembly
memcpy.s and excludes the C.  So "i386 unchanged" here is true by
construction rather than by care.

+++ uros/src/lib/libmach/bcopy.c
@@ -88,8 +88,19 @@
 /*
  * sizeof(word) MUST BE A POWER OF TWO
  * SO THAT wmask BELOW IS ALL ONES
+ *
+ * ⚠️ As wide as a pointer, which is the whole point of the type and is what
+ * "optimal copy speed" meant when it was written.  It said `int', and on the
+ * machine it was written for that was the same thing; on x86-64 it is half.
+ * Every memcpy, memmove and bcopy in user space was moving 32 bits per
+ * iteration on a 64-bit machine, and aligning its preamble to four bytes
+ * rather than eight.
+ *
+ * uintptr_t rather than `long' so the intent is stated rather than inferred
+ * from a type that happens to be right on both targets.  On i386 it is four
+ * bytes, exactly as before, which is why i386 does not move.
  */
-typedef	int word;		/* "word" used for optimal copy speed */
+typedef	uintptr_t word;		/* "word" used for optimal copy speed */
 
 #define	wsize	sizeof(word)
 #define	wmask	(wsize - 1)
@@ -128,13 +139,21 @@ bcopy(const void *src0, void *dst0, size_t length)
 		/*
 		 * Copy forward.
 		 */
-		t = (int)src;	/* only need low bits */
-		if ((t | (int)dst) & wmask) {
+		/*
+		 * ⚠️ uintptr_t, not int.  Only the low bits are read -- every
+		 * use below is masked with wmask -- so truncating to int was
+		 * harmless, and that is exactly what made it worth removing:
+		 * a diagnosis that is correct and harmless is the kind that
+		 * teaches a reader to skip the class, and this class had a
+		 * real one in it.
+		 */
+		t = (uintptr_t)src;	/* only need low bits */
+		if ((t | (uintptr_t)dst) & wmask) {
 			/*
 			 * Try to align operands.  This cannot be done
 			 * unless the low bits match.
 			 */
-			if ((t ^ (int)dst) & wmask || length < wsize)
+			if ((t ^ (uintptr_t)dst) & wmask || length < wsize)
 				t = length;
 			else
 				t = wsize - (t & wmask);
@@ -157,9 +176,9 @@ bcopy(const void *src0, void *dst0, size_t length)
 		 */
 		src += length;
 		dst += length;
-		t = (int)src;
-		if ((t | (int)dst) & wmask) {
-			if ((t ^ (int)dst) & wmask || length <= wsize)
+		t = (uintptr_t)src;
+		if ((t | (uintptr_t)dst) & wmask) {
+			if ((t ^ (uintptr_t)dst) & wmask || length <= wsize)
 				t = length;
 			else
 				t &= wmask;
